Truck Log Managment/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
requests==2.32.3
